# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Red Solidaria San Ramón API
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export type CampaignStatus =
  (typeof CampaignStatus)[keyof typeof CampaignStatus];

export const CampaignStatus = {
  active: "active",
  completed: "completed",
  paused: "paused",
} as const;

export interface Campaign {
  id: number;
  title: string;
  description: string;
  imageUrl?: string | null;
  goal: number;
  raised: number;
  donorCount: number;
  status: CampaignStatus;
  featured: boolean;
  category: string;
  startDate: string;
  endDate?: string | null;
  createdAt: string;
}

export type CreateCampaignInputStatus =
  (typeof CreateCampaignInputStatus)[keyof typeof CreateCampaignInputStatus];

export const CreateCampaignInputStatus = {
  active: "active",
  completed: "completed",
  paused: "paused",
} as const;

export interface CreateCampaignInput {
  title: string;
  description: string;
  imageUrl?: string | null;
  goal: number;
  raised?: number;
  status: CreateCampaignInputStatus;
  featured: boolean;
  category: string;
  startDate: string;
  endDate?: string | null;
}

export interface CampaignUpdate {
  id: number;
  campaignId: number;
  title: string;
  content: string;
  createdAt: string;
}

export interface CreateCampaignUpdateInput {
  title: string;
  content: string;
}

export interface CampaignImage {
  id: number;
  campaignId: number;
  imageUrl: string;
  caption?: string | null;
  createdAt: string;
}

export interface CreateCampaignImageInput {
  imageUrl: string;
  caption?: string | null;
}

export type DonationStatus =
  (typeof DonationStatus)[keyof typeof DonationStatus];

export const DonationStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export interface Donation {
  id: number;
  campaignId?: number | null;
  campaignTitle?: string | null;
  firstName: string;
  lastName: string;
  email: string;
  phone?: string | null;
  amount: number;
  paymentMethod: string;
  message?: string | null;
  anonymous: boolean;
  publicProof?: boolean;
  receiptUrl?: string | null;
  receiptNote?: string | null;
  status: DonationStatus;
  adminNote?: string | null;
  createdAt: string;
}

export type CreateDonationInputPaymentMethod =
  (typeof CreateDonationInputPaymentMethod)[keyof typeof CreateDonationInputPaymentMethod];

export const CreateDonationInputPaymentMethod = {
  yape: "yape",
  plin: "plin",
  transfer: "transfer",
  card: "card",
  cash: "cash",
  other: "other",
} as const;

export interface CreateDonationInput {
  campaignId?: number | null;
  firstName: string;
  lastName: string;
  email: string;
  phone?: string | null;
  amount: number;
  paymentMethod: CreateDonationInputPaymentMethod;
  message?: string | null;
  anonymous: boolean;
  publicProof?: boolean;
  receiptUrl?: string | null;
  receiptNote?: string | null;
  proofImageUrl?: string | null;
  proofPublicId?: string | null;
  proofMimeType?: string | null;
}

export interface CampaignDonor {
  id: number;
  name?: string | null;
  amount: number;
  message?: string | null;
  date: string;
  publicProof: boolean;
  proofUrl?: string | null;
}

export interface DonationProof {
  id: number;
  donationId: number;
  imageUrl: string;
  publicId?: string | null;
  mimeType?: string | null;
  createdAt: string;
}

export interface UploadSignature {
  cloudName: string;
  apiKey: string;
  timestamp: string;
  publicId: string;
  folder: string;
  signature: string;
}

export type UpdateDonationStatusInputStatus =
  (typeof UpdateDonationStatusInputStatus)[keyof typeof UpdateDonationStatusInputStatus];

export const UpdateDonationStatusInputStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export interface UpdateDonationStatusInput {
  status: UpdateDonationStatusInputStatus;
  adminNote?: string | null;
}

export interface DonationStats {
  totalDonations: number;
  totalAmount: number;
  pendingCount: number;
  approvedCount: number;
  totalDonors: number;
}

export interface NewsPost {
  id: number;
  title: string;
  content: string;
  summary: string;
  imageUrl?: string | null;
  publishedAt: string;
  createdAt: string;
}

export interface CreateNewsInput {
  title: string;
  content: string;
  summary: string;
  imageUrl?: string | null;
  publishedAt: string;
}

export interface Testimonial {
  id: number;
  name: string;
  role: string;
  message: string;
  avatarUrl?: string | null;
  createdAt: string;
}

export interface CreateTestimonialInput {
  name: string;
  role: string;
  message: string;
  avatarUrl?: string | null;
}

export interface ImpactStats {
  childrenHelped: number;
  campaignsRun: number;
  volunteers: number;
  donationsReceived: number;
  animalsHelped: number;
}

export interface ContactMessageInput {
  name: string;
  email: string;
  phone?: string | null;
  message: string;
  subject: string;
}

export interface ContactMessage {
  id: number;
  name: string;
  email: string;
  phone?: string | null;
  message: string;
  subject: string;
  createdAt: string;
}

export interface VolunteerInput {
  name: string;
  email: string;
  phone?: string | null;
  availability: string;
  skills?: string | null;
  motivation?: string | null;
}

export interface Volunteer {
  id: number;
  name: string;
  email: string;
  phone?: string | null;
  availability: string;
  skills?: string | null;
  motivation?: string | null;
  createdAt: string;
}

export interface AdminLoginInput {
  username: string;
  password: string;
}

export interface AdminUser {
  id: number;
  username: string;
}

export interface AdminLoginResponse {
  success: boolean;
  message: string;
  user?: AdminUser;
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export type GetCampaignsParams = {
  status?: GetCampaignsStatus;
  featured?: boolean;
};

export type GetCampaignsStatus =
  (typeof GetCampaignsStatus)[keyof typeof GetCampaignsStatus];

export const GetCampaignsStatus = {
  active: "active",
  completed: "completed",
  paused: "paused",
  all: "all",
} as const;

export type GetDonationsParams = {
  campaignId?: number;
  status?: GetDonationsStatus;
};

export type GetDonationsStatus =
  (typeof GetDonationsStatus)[keyof typeof GetDonationsStatus];

export const GetDonationsStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export type AddDonationProofBody = {
  imageUrl: string;
  publicId?: string | null;
  mimeType?: string | null;
};