# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Red Solidaria San Ramón API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /campaigns:
    get:
      operationId: getCampaigns
      tags: [campaigns]
      summary: List all campaigns
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, completed, paused, all]
        - name: featured
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of campaigns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Campaign"
    post:
      operationId: createCampaign
      tags: [campaigns]
      summary: Create a campaign (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCampaignInput"
      responses:
        "201":
          description: Created campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{id}:
    get:
      operationId: getCampaign
      tags: [campaigns]
      summary: Get campaign by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
    put:
      operationId: updateCampaign
      tags: [campaigns]
      summary: Update campaign (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCampaignInput"
      responses:
        "200":
          description: Updated campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
    delete:
      operationId: deleteCampaign
      tags: [campaigns]
      summary: Delete campaign (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /campaigns/{id}/donors:
    get:
      operationId: getCampaignDonors
      tags: [donations]
      summary: Public list of approved donors for a campaign (no PII)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of approved donors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CampaignDonor"
...
components:
  schemas:
...
    Campaign:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        imageUrl:
          type: string
          nullable: true
        goal:
          type: number
        raised:
          type: number
        donorCount:
          type: integer
        status:
          type: string
          enum: [active, completed, paused]
        featured:
          type: boolean
        category:
          type: string
        startDate:
          type: string
        endDate:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - title
        - description
        - goal
        - raised
        - donorCount
        - status
        - featured
        - category
        - startDate
        - createdAt
    CreateCampaignInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        imageUrl:
          type: string
          nullable: true
        goal:
          type: number
        raised:
          type: number
        status:
          type: string
          enum: [active, completed, paused]
        featured:
          type: boolean
        category:
          type: string
        startDate:
          type: string
        endDate:
          type: string
          nullable: true
      required:
        - title
        - description
        - goal
        - status
        - featured
        - category
        - startDate
...
    CampaignDonor:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
          nullable: true
        amount:
          type: number
        message:
          type: string
          nullable: true
        date:
          type: string
        publicProof:
          type: boolean
        proofUrl:
          type: string
          nullable: true
      required:
        - id
        - amount
        - date
        - publicProof
...