# lib/api-zod/src/generated/types/campaignDonor.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Red Solidaria San Ramón API
 * OpenAPI spec version: 0.2.0
 */

export interface CampaignDonor {
  id: number;
  name?: string | null;
  amount: number;
  message?: string | null;
  date: string;
  publicProof: boolean;
  proofUrl?: string | null;
}